refactor: split repositoryInformation function from fork
