+++ source-code/website/src/services/github/index.ts
@@ -1,5 +1,3 @@
 export { isCollaborator } from "./isCollaborator.telefunc.js";
-export { onFork } from "./fork.telefunc.js";
-export { isFork } from "./fork.telefunc.js";
-
-export { syncFork } from "./fork.telefunc.js";
+export { onFork, syncFork } from "./fork.telefunc.js";
+export { repositoryInformation } from "./repositoryInformation.telefunc.js";
